pyproject.toml
[tool.poetry]
name = "cakto-challenge"
version = "0.1.0"
description = "Mini Split Engine + Ledger + Outbox - Teste Backend Sênior Cakto"
authors = ["Gabriel Rodrigues"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
gunicorn = "^23.0"
django-injector = "^0.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.9"
pytest-cov = "^6.0"
ruff = "^0.11"
black = "^25.1"
factory-boy = "^3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# --- Pytest ---
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.settings"
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

# --- Black ---
[tool.black]
line-length = 120
target-version = ["py311"]

# --- Ruff ---
[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
